pitch-system-api/pyproject.toml
[tool.poetry]
name = "pitch-system-api"
version = "0.1.0"
description = ""
authors = ["Aonrud <aonrud@fastmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
djangorestframework = "^3.15.2"
mysqlclient = "^2.2.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
